moved touchable profile to genral components folder
